requirements.txt
pandas==2.1.2
requests==2.31.0
gspread==5.12.0
statistics==1.0.3.5
boto3==1.29.6
botocore==1.32.6
pytz==2023.3.post1
cachetools==5.3.2
google-api-python-client==2.108.0
oauth2client==4.1.3
rapidfuzz==3.5.2
bs4==0.0.1
streamlit
pytz
arrow
Pillow
geocoder